New format of client message, and Server Processing
Updated the message that the client sends to the sever. Now has a short
form direction and timestamp.

Server now can process the timestamp and save the timestamp of the last
message the player sent to the server.

Also added a bool flag to the paddle to check when the timestamp has
changed. fun times.

TO ADD
- Calculating latency